Complete the Add Venue page and the ability to use the page to add a new Venue to the database
Add the ability to create and inset a new venue to the database from
the New Venue page. Create a modified AnyOf enum restriction validator
for the SelectMultipleFields form element that checks if the values
selected from the form is a subset of the allowed choices. Apply
CSRF token for form security.