Add functionality to create shows
Add the ability to allow user to create a show, with an artist, a
venue, and starting time. Modify the show form class to show a selection
of the available artists and venues. Modify the show form init method
to update automatically whenever a new venue or artist is created.